# backend/docs/swagger.yaml
definitions:
  internal_handlers.AuthResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user: {}
    type: object
...
  internal_handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  internal_handlers.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  internal_handlers.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  internal_handlers.UpdateEstablishmentRequest:
    properties:
      active:
        type: boolean
      address:
        type: string
      email:
        type: string
      has_seating_places:
        type: boolean
      name:
        type: string
      phone:
        type: string
      table_count:
        type: integer
      type:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Аутентифицирует пользователя и возвращает токены доступа
      parameters:
      - description: Данные для входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Вход в систему
      tags:
      - auth
  /auth/logout:
    post:
      description: Добавляет токен доступа в черный список
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Выход из системы
      tags:
      - auth
...
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Обновляет access token используя refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновление токена
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Регистрирует нового пользователя и возвращает токены доступа
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handlers.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Регистрация нового пользователя
      tags:
      - auth
  /establishments:
    get:
      description: Возвращает заведения пользователя (обычно 0 или 1 элемент)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Получить список заведений
      tags:
      - establishments
    post:
      consumes:
      - application/json
      description: Заведение создаётся при прохождении onboarding, этот endpoint не
        используется
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Создать заведение (не используется)
      tags:
      - establishments
  /establishments/{id}:
    delete:
      description: Удаление заведения не разрешено
      parameters:
      - description: ID заведения
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Удалить заведение (не разрешено)
      tags:
      - establishments
    get:
      description: Возвращает заведение по ID (доступ только к своему заведению)
      parameters:
      - description: ID заведения
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Получить заведение по ID
      tags:
      - establishments
    put:
      consumes:
      - application/json
      description: Обновляет заведение (доступ только к своему заведению)
      parameters:
      - description: ID заведения
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.UpdateEstablishmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Обновить заведение
      tags:
      - establishments
  /finance/cash-flow:
    get:
      description: Возвращает отчет о движении денежных средств
      parameters:
      - description: Начальная дата
        in: query
        name: start_date
        type: string
      - description: Конечная дата
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Получить отчет о движении денежных средств
      tags:
      - finance
  /finance/pnl:
    get:
      description: Возвращает отчет о прибылях и убытках
      parameters:
      - description: Начальная дата
        in: query
        name: start_date
        type: string
      - description: Конечная дата
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Получить отчет P&L
      tags:
      - finance
  /finance/shifts:
    get:
      description: Возвращает список смен
      parameters:
      - description: Начальная дата
        in: query
        name: start_date
        type: string
      - description: Конечная дата
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Получить смены
      tags:
      - finance
  /finance/transactions:
    get:
      description: Возвращает список финансовых транзакций
      parameters:
      - description: Начальная дата
        in: query
        name: start_date
        type: string
      - description: Конечная дата
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Получить транзакции
      tags:
      - finance
...
swagger: "2.0"