# backend/internal/handlers/finance_handler.go
package handlers

import (
	"net/http"

	"github.com/gin-gonic/gin"
	"go.uber.org/zap"

	"github.com/yourusername/arc/backend/internal/usecases"
)

type FinanceHandler struct {
	usecase *usecases.FinanceUseCase
	logger  *zap.Logger
}

func NewFinanceHandler(usecase *usecases.FinanceUseCase, logger *zap.Logger) *FinanceHandler {
	return &FinanceHandler{
		usecase: usecase,
		logger:  logger,
	}
}

// GetTransactions возвращает транзакции
// @Summary Получить транзакции
// @Description Возвращает список финансовых транзакций
// @Tags finance
// @Produce json
// @Security Bearer
// @Param start_date query string false "Начальная дата"
// @Param end_date query string false "Конечная дата"
// @Success 200 {object} map[string]interface{}
// @Router /finance/transactions [get]
func (h *FinanceHandler) GetTransactions(c *gin.Context) {
	// TODO: Implement get transactions logic
	c.JSON(http.StatusOK, gin.H{"data": []interface{}{}})
}

// GetShifts возвращает смены
// @Summary Получить смены
// @Description Возвращает список смен
// @Tags finance
// @Produce json
// @Security Bearer
// @Param start_date query string false "Начальная дата"
// @Param end_date query string false "Конечная дата"
// @Success 200 {object} map[string]interface{}
// @Router /finance/shifts [get]
func (h *FinanceHandler) GetShifts(c *gin.Context) {
	// TODO: Implement get shifts logic
	c.JSON(http.StatusOK, gin.H{"data": []interface{}{}})
}

// GetPNL возвращает отчет о прибылях и убытках
// @Summary Получить отчет P&L
// @Description Возвращает отчет о прибылях и убытках
// @Tags finance
// @Produce json
// @Security Bearer
// @Param start_date query string false "Начальная дата"
// @Param end_date query string false "Конечная дата"
// @Success 200 {object} map[string]interface{}
// @Router /finance/pnl [get]
func (h *FinanceHandler) GetPNL(c *gin.Context) {
	// TODO: Implement P&L report logic
	c.JSON(http.StatusOK, gin.H{"data": map[string]interface{}{}})
}

// GetCashFlow возвращает отчет о движении денежных средств
// @Summary Получить отчет о движении денежных средств
// @Description Возвращает отчет о движении денежных средств
// @Tags finance
// @Produce json
// @Security Bearer
// @Param start_date query string false "Начальная дата"
// @Param end_date query string false "Конечная дата"
// @Success 200 {object} map[string]interface{}
// @Router /finance/cash-flow [get]
func (h *FinanceHandler) GetCashFlow(c *gin.Context) {
	// TODO: Implement cash flow report logic
	c.JSON(http.StatusOK, gin.H{"data": map[string]interface{}{}})
}